+++ minion/parallel.cpp
@@ -6,7 +6,7 @@
 #include "parallel/parallel.h"
 
 // Disable on windows
-#ifndef _WIN32
+#if !defined(_WIN32) && !defined(__EMSCRIPTEN__)
 #define PARALLEL
 #endif